Refactor into modules
for dropbox, photos, and storage

index.js is just a simple entry point at this point.
